requirements.txt
streamlit
torch
torchvision
Pillow
numpy
matplotlib
grad-cam